Working Flask API for writing to Ditto.db sqlite3 database. Ready for ditto-app client.
