--- gui.py
@@ -1,16 +1,0 @@
-'''
-GUI for the assistant.
-
-author: Omar Barazanji
-
-sources:
-1) https://build-system.fman.io/pyqt5-tutorial
-'''
-
-# from PyQt5.QtWidgets import QApplication, QLabel
-
-# if __name__ == "__main__":
-#     app = QApplication([])
-#     label = QLabel('This is a test.')
-#     label.show()
-#     app.exec()
